reformat popups
